feat: refactor BOM generation into sub-command `make-bom`
BREAKING CHANGE
